+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # NEXT UP
 
 - Legendary resistance effect was no longer working properly.
-
+- Small tweak to detecting primary spellcasting ability, that now defaults to the value set by the subclass rather than class. This is used by some custom subclasses that might change the default value. @sayshal
 
 # 3.4.62